Windows: json-c 0.17 compatibility with ssize_t type definition (part 2)
Additional changes required for Visual Studio 0.103 build.

+++ win32/clamav-config.h
@@ -573,7 +573,8 @@
 /* #undef off_t */
 
 /* Define to `int' */
-typedef int ssize_t;
+#include <BaseTsd.h>
+typedef SSIZE_T ssize_t;
 
 /* Define to the equivalent of the C99 'restrict' keyword, or to
    nothing if this is not supported.  Do not define if restrict is
